requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
cryptography==43.0.0
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==5.0.7
django-allauth==64.0.0
django-cache-url==3.4.5
django-crispy-forms==2.3
django-jalali-date==1.1.3
django-rosetta==0.10.0
environs==11.0.0
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
marshmallow==3.21.3
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
polib==1.2.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
